+++ test/transform/resource/before/BuilderDefaults.java
@@ -1,0 +1,9 @@
+import lombok.Builder;
+import lombok.Value;
+
+@Value @Builder
+public class BuilderDefaults {
+	@Builder.Default int x = 10;
+	String name;
+	@Builder.Default long z = System.currentTimeMillis();
+}